requirements.txt
paho-mqtt~=1.6.1
APScheduler~=3.9.1
requests~=2.27.1
actions_server~=0.0.6
service-bootstrap~=0.0.1
